brain/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyyaml==6.0.2
redis==5.2.0
face-recognition==1.3.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
python-multipart==0.0.12